Conflicts:
	PokTaPok/makefile
	PokTaPok/src/poktapokAgent.cpp
	PokTaPok/src/poktapokAgent.h

+++ PokTaPok/src/localizationMethods.h
@@ -9,5 +9,5 @@ Vector2D ubicacionBanderaBandera(double distanciaFlag1, double direccionFlag1, V
 Vector2D ubicacionLineaBandera(double direccionLinea, char lineaId,double distanciaFlag, double direccionFlag, Vector2D coordGlobalFlag );
 double  orientacion_global(Vector2D coordGlobalFlag,Vector2D coordGlobalAgente, double angulo_relativo );
 double  body_angle_loc(int head_angle, Vector2D global_agent,  Vector2D coordGlobalFlag, int sensed_flag_direction );
-int orientation(char a);
+double orientation(char a);
 #endif
